refactor: 디자인 1차 수정
