+++ src/components/common/selectable-chip/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Chip } from '@heroui/react';
+import Chip from '../chip';
 
 interface Props {
   label: string;
@@ -11,11 +11,7 @@ interface Props {
 export default function ClickableChip({ label, isSelected, onClick }: Props) {
   return (
     <button type="button" onClick={onClick} className="focus:outline-none">
-      <Chip
-        className={`cursor-pointer ${
-          isSelected ? 'bg-blue-600 text-white hover:bg-blue-700' : 'bg-gray-100 text-gray-800 hover:bg-gray-200'
-        }`}
-      >
+      <Chip size="lg" className={`cursor-pointer ${isSelected ? 'bg-blue-600 text-white hover:bg-blue-700' : ''}`}>
         {label}
       </Chip>
     </button>
